requirements.txt
decorator==5.1.1
wheel==0.37.1
appdirs==1.4.4
arrow==1.2.2
backports-abc==0.5
certifi==2021.10.8
chardet==4.0.0
idna==3.3
praw==7.5.0
prawcore==2.3.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
six==1.16.0
tornado==6.1
update-checker==0.18.0
urllib3==1.26.8
testresources==2.0.1
